refactor: remove barrel
